[refactor] : 패키지 이동, 클래스명 변경
